backend/mobile/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
